Move nordigen to separate directory
